Typecast uint64_t to long long unsigned while printing

+++ runtime/legion/legion_profiling_serializer.cc
@@ -1661,7 +1661,8 @@ namespace Legion {
     {
       log_prof.print("Machine Desc %d %d %s %llu %d",
                      machine_desc.node_id, machine_desc.num_nodes,
-                     machine_desc.process_info.hostname, machine_desc.process_info.hostid,
+                     machine_desc.process_info.hostname,
+                     (unsigned long long)machine_desc.process_info.hostid,
                      machine_desc.process_info.processid);
     }
